Refactor encounter types and add Day1 encounters
Moved Day0 encounter scripts into a dedicated Day0 folder and adjusted their health values from 69 to 20. Added new encounter types for Day1, including AgileDvorniashka, BlackDvorniashka, BouncyDvorniashka, Dvorniashka, HotDvorniashka, RedDvorniashka, and SpikyDvorniashka, each with their own initialization logic. Updated Encounter.cs to use float for goldRewardMod and its getter. Minor update to Project Overview(Durak).docx.

+++ Assets/Scripts/Encounter/Encounter.cs
@@ -9,7 +9,7 @@ public abstract class Encounter : ScriptableObject
     [SerializeField] protected int day;
     [SerializeField] protected List<CardInfo> deck;
     [SerializeField] protected int health;
-    [SerializeField] protected int goldRewardMod;
+    [SerializeField] protected float goldRewardMod;
     [SerializeField] protected char trumpSuit;
     [SerializeField] protected  Sprite icon;
     [SerializeField] protected bool boss;
@@ -24,7 +24,7 @@ public abstract class Encounter : ScriptableObject
     {
         return deck;
     }
-    public int GetReward() 
+    public float GetReward() 
     {
         return goldRewardMod;
     }
